feat: adcionado api completa, validators filtros e regras de negocio

+++ CursoDeIdiomas.Application/Services/TurmaService.cs
@@ -1,9 +1,12 @@
 ﻿using AutoMapper;
+using CursoDeIdiomas.Application.Exceptions;
 using CursoDeIdiomas.Application.Interfaces;
 using CursoDeIdiomas.Application.Params;
+using CursoDeIdiomas.Application.ViewModels.Aluno;
 using CursoDeIdiomas.Application.ViewModels.Turma;
 using CursoDeIdiomas.Domain.Interfaces;
 using CursoDeIdiomas.Domain.Models;
+using FluentValidation;
 using Microsoft.EntityFrameworkCore;
 
 namespace CursoDeIdiomas.Application.Services
@@ -13,14 +16,22 @@ namespace CursoDeIdiomas.Application.Services
         private readonly ITurmaRepository _turmaRepository;
         private readonly IMapper _mapper;
         private readonly IUnitOfWork _unityOfWork;
+        private readonly IValidator<TurmaRequest> _validator;
 
-        public TurmaService(ITurmaRepository turmaRepository, IMapper mapper, IUnitOfWork unityOfWork)
+        public TurmaService(
+            ITurmaRepository turmaRepository, 
+            IMapper mapper, 
+            IUnitOfWork unityOfWork,
+            IValidator<TurmaRequest> validator
+            )
         {
             _turmaRepository = turmaRepository;
             _mapper = mapper;
             _unityOfWork = unityOfWork;
+            _validator = validator;
 
             _turmaRepository.AddPreQuery(x => x.Include(x => x.Alunos));
+            _turmaRepository.AddPreQuery(x => x.Include(x => x.Curso));
         }
 
         public async Task<IEnumerable<TurmaResponse>> GetAsync(TurmaParams queryParams)
@@ -35,9 +46,9 @@ namespace CursoDeIdiomas.Application.Services
 
         public async Task<TurmaResponse> AddAsync(TurmaRequest turmaRequest)
         {
-            //var validation = await _validator.ValidateAsync(contactRequest);
-            //if (!validation.IsValid)
-            //throw new BadRequestException(validation);
+            var validation = await _validator.ValidateAsync(turmaRequest);
+            if (!validation.IsValid)
+            throw new BadRequestException(validation);
 
             var turma = _mapper.Map<Turma>(turmaRequest);
 
@@ -54,9 +65,9 @@ namespace CursoDeIdiomas.Application.Services
                 throw new Exception($"Turma com id: {id} não existe! ");
 
             _turmaRepository.AddPreQuery(x => x.Where(c => c.Id != id));
-            //var validation = await _validator.ValidateAsync(alunoRequest);
-            //if (!validation.IsValid)
-            //throw new BadRequestException(validation);
+            var validation = await _validator.ValidateAsync(turmaRequest);
+            if (!validation.IsValid)
+            throw new BadRequestException(validation);
 
             _mapper.Map<TurmaRequest, Turma>(turmaRequest, existing);
             await _turmaRepository.UpdateAsync(existing);
@@ -69,6 +80,9 @@ namespace CursoDeIdiomas.Application.Services
             var existing = await _turmaRepository.GetByIdAsync(id);
             if (existing is null)
                 throw new Exception($"Turma com id: {id} não existe! ");
+
+            if(existing.Alunos.Any())
+                throw new Exception($"Turma não pode ser excluída se possuir alunos.");
 
             await _turmaRepository.DeleteAsync(id);
             await _unityOfWork.CommitAsync();
